Update Unity grammars, and change URLs in text.
The VOUnits.tex document was pointing to bitbucket.org and Google Code;
I've updated both of these to point to appropriate new URLs.

The file known-units.tex is generated from unity-grammars.zip, which is
in turn copied from the Unity repository.  This is from
revision 1037bf0656bd (a snapshot rather than a formal release), of
https://hg.sr.ht/~nxg/unity.  The outstanding library fixes have been
done, but I should wait until the VOUnits revision is further advanced,
before releasing v1.1.