Add text to discuss scale-factors in function-application. (#22)
Addresses issue #17.

Despite the discussion in that issue, the preferred grammars (or rather,
the grammars as (not) adjusted in relation to [unity issue
not permit scalefactors.

Firstly, there's only ambiguous justification for including them, based on the described FITS/OGIP grammars.  Secondly, the fact that FITS permits `m(2)`, meaning `m^2`, means that function-application is nearly ambiguous anyway, and I was not able to find a way of resolving that in a straightforward way.  I think we can take that as a hint that we shouldn't be creative here.