Reworded the note commenting on the bessellian/tropical year deprecation.
This matches the reversion in Unity repo
<https://heptapod.host/nxg/unity/> revision 646f54a6edb4, which fixed
the earlier revision (b45b533ada39) which incorrectly marked these as
not permitted, rater than deprecated.

Note that, since Table 2 is generated from a file in the Unity
distribution, which will be added here when Unity v1.1 is released,
the table as it appears in this document revision still incorrectly
shows these units as forbidden.